Modified lvl1
The search content for lvl1 was null as the h1 tag is present under the article tag and not under header. Due to absence of lvl1 the desired search results were not shown for the given keyword. For example, if the search term is 'Docker' then the expected result is that the page with h1 tag as Docker should be given precedence. However, as lvl1 was null the precedence was given to lvl2 and the page was not shown in the search results at the designated position.